Added visualization within text-based version of the game to see BFS algorithm executed.
BFS algorithm also adds adjacent tiles in a new order. Previously, this order was north, east, south, west. Not, the order is east, north, south, west. This is because of the generated position and texture of the starting tile appearing to hint that the target tile will be east of it.